refactor: move functions into their files
